requirements.txt
# 核心框架
langchain-core>=1.0.0,<2.0.0
# 1.x 主包（2025-05 发布的 1.0 正式版）
langchain>=1.0.0,<2.0.0
# 各模型实现（同步升到 1.x）
langchain-openai>=1.0.0,<2.0.0
langchain-community>=0.3.31,<1.0
langchain-anthropic>=1.0.0,<2.0.0

# Web框架
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
jinja2==3.1.2

# 数据库和缓存
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
celery==5.3.4

# 任务队列和异步
asyncio-mqtt==0.16.1
aiofiles==23.2.1
aioredis==2.0.1

# 浏览器自动化
selenium==4.15.2
playwright==1.40.0

# MCP集成
# mcp==0.1.0
httpx==0.27.0

# 数据处理
pydantic==2.7.4
pyyaml==6.0.1
python-dotenv==1.0.0

# 日志和监控
loguru==0.7.2
prometheus-client==0.19.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# 开发工具
black==23.11.0
flake8==6.1.0
mypy==1.7.1